backend/requirements.txt
fastapi
uvicorn
statistics
matplotlib
aiohttp
jinja2
python-dotenv
pytest
pytest-asyncio